+++ source/updateCheck.py
@@ -67,7 +67,7 @@ import winKernel
 from utils.tempFile import _createEmptyTempFileForDeletingFile
 
 #: The URL to use for update checks.
-CHECK_URL = "https://www.nvaccess.org/nvdaUpdateCheck"
+_DEFAULT_CHECK_URL = "https://www.nvaccess.org/nvdaUpdateCheck"
 #: The time to wait between checks.
 CHECK_INTERVAL = 86400  # 1 day
 #: The time to wait before retrying a failed check.
@@ -89,6 +89,12 @@ state: Optional[Dict[str, Any]] = None
 #: The single instance of L{AutoUpdateChecker} if automatic update checking is enabled,
 #: C{None} if it is disabled.
 autoChecker: Optional["AutoUpdateChecker"] = None
+
+
+def _getCheckURL() -> str:
+	if url := config.conf["update"]["serverURL"]:
+		return url
+	return _DEFAULT_CHECK_URL
 
 
 def getQualifiedDriverClassNameForStats(cls):
@@ -162,7 +168,7 @@ def checkForUpdate(auto: bool = False) -> Optional[Dict]:
 			"outputBrailleTable": config.conf["braille"]["translationTable"] if brailleDisplayClass else None,
 		}
 		params.update(extraParams)
-	url = "%s?%s" % (CHECK_URL, urllib.parse.urlencode(params))
+	url = f"{_getCheckURL()}?{urllib.parse.urlencode(params)}"
 	try:
 		log.debug(f"Fetching update data from {url}")
 		res = urllib.request.urlopen(url, timeout=UPDATE_FETCH_TIMEOUT_S)
@@ -331,12 +337,31 @@ class UpdateChecker(garbageHandler.TrackedObject):
 		)
 
 	def _error(self):
+		if url := config.conf["update"]["serverURL"]:
+			tip = pgettext(
+				"updateCheck",
+				# Translators: A suggestion of what to do when checking for NVDA updates fails and an update mirror is being used.
+				# {url} will be replaced with the mirror URL.
+				"Make sure you are connected to the internet, and the NVDA update mirror URL is valid.\n"
+				"Mirror URL: {url}",
+			).format(url=url)
+		else:
+			tip = pgettext(
+				"updateCheck",
+				# Translators: A suggestion of what to do when fetching add-on data from the store fails and the default metadata URL is being used.
+				"Make sure you are connected to the internet and try again.",
+			)
+		message = pgettext(
+			"updateCheck",
+			# Translators: A message indicating that an error occurred while checking for an update to NVDA.
+			# tip will be replaced with a context sensitive suggestion of next steps.
+			"Error checking for update.\n{tip}",
+		).format(tip=tip)
 		wx.CallAfter(self._progressDialog.done)
 		self._progressDialog = None
 		wx.CallAfter(
 			gui.messageBox,
-			# Translators: A message indicating that an error occurred while checking for an update to NVDA.
-			_("Error checking for update."),
+			message,
 			# Translators: The title of an error message dialog.
 			_("Error"),
 			wx.OK | wx.ICON_ERROR,
@@ -943,7 +968,7 @@ def _updateWindowsRootCertificates():
 	with requests.get(
 		# We must specify versionType so the server doesn't return a 404 error and
 		# thus cause an exception.
-		CHECK_URL + "?versionType=stable",
+		f"{_getCheckURL()}?versionType=stable",
 		timeout=UPDATE_FETCH_TIMEOUT_S,
 		# Use an unverified connection to avoid a certificate error.
 		verify=False,